Project Manager - Can configure guidelines from PM SettingsView
