+++ avail/src/main/kotlin/avail/anvil/settings/SettingsView.kt
@@ -32,35 +32,18 @@
 
 package avail.anvil.settings
 
-import avail.anvil.environment.availStandardLibraries
-import avail.anvil.icons.ProjectManagerIcons
 import avail.anvil.manager.AvailProjectManager
-import avail.anvil.shortcuts.KeyboardShortcutCategory
-import avail.anvil.versions.MavenCentralAPI
-import org.availlang.artifact.environment.AvailEnvironment
 import java.awt.Color
 import java.awt.Dialog.ModalityType.DOCUMENT_MODAL
 import java.awt.Dimension
-import java.awt.FlowLayout
 import java.awt.Font
-import java.awt.GridBagConstraints
-import java.awt.GridBagLayout
-import java.awt.event.MouseAdapter
-import java.awt.event.MouseEvent
-import java.io.File
 import javax.swing.BorderFactory
-import javax.swing.Box
 import javax.swing.BoxLayout
-import javax.swing.JButton
 import javax.swing.JDialog
 import javax.swing.JFrame
-import javax.swing.JLabel
-import javax.swing.JOptionPane
 import javax.swing.JPanel
 import javax.swing.JScrollPane
-import javax.swing.JTabbedPane
 import javax.swing.ScrollPaneConstants
-import javax.swing.SwingUtilities
 
 /**
  * A [JFrame] that displays configurable global settings.
@@ -128,349 +111,3 @@ class SettingsView constructor (
 	}
 }
 
-/**
- * A [JPanel] that represents a settings category.
- *
- * @author Richard Arriaga
- *
- * @property labelText
- *   The text to show in the panel label.
- * @param settingsView
- *   The parent [SettingsView].
- */
-sealed class SettingPanelSelection constructor(
-	private val labelText: String,
-	protected val settingsView: SettingsView
-):JPanel(FlowLayout(FlowLayout.LEFT))
-{
-	/**
-	 * The [JLabel] that displays the [labelText].
-	 */
-	val label =  JLabel(labelText)
-
-	fun init()
-	{
-		minimumSize = Dimension(150, 35)
-		preferredSize = Dimension(150, 35)
-		maximumSize = Dimension(150, 35)
-		addMouseListener(object: MouseAdapter()
-		{
-			override fun mouseClicked(e: MouseEvent)
-			{
-				if (e.clickCount == 1)
-				{
-					select()
-				}
-			}
-		})
-		add(label)
-	}
-
-	/**
-	 * Update the [settingsView] right panel with the appropriate settings
-	 * configuration screen associated with this [SettingPanelSelection].
-	 */
-	abstract fun updateSettingsPane ()
-
-	/**
-	 * Select this [SettingPanelSelection].
-	 */
-	fun select ()
-	{
-		if (this == settingsView.selected) return
-		label.font = label.font.deriveFont(font.style or Font.BOLD)
-		background = Color(0x55, 0x58, 0x5A)
-		settingsView.selected.deselect()
-		settingsView.selected = this
-		updateSettingsPane()
-	}
-
-	/**
-	 * Deselect this [SettingPanelSelection].
-	 */
-	private fun deselect ()
-	{
-		label.font = label.font.deriveFont(font.style or Font.PLAIN)
-		background = settingsView.background
-	}
-}
-
-/**
- * The [SettingPanelSelection] used for showing the shortcut settings.
- *
- * @author Richard Arriaga
- *
- * @constructor
- * Construct a [ShortcutsSelection].
- *
- * @param settingsView
- *   The parent [SettingsView].
- */
-class ShortcutsSelection constructor(
-	settingsView: SettingsView
-): SettingPanelSelection("Shortcuts", settingsView)
-{
-	override fun updateSettingsPane()
-	{
-		val tabs = JTabbedPane(JTabbedPane.TOP, JTabbedPane.WRAP_TAB_LAYOUT)
-
-		KeyboardShortcutCategory.values().forEach {
-			tabs.addTab(
-				it.display,
-				ShortcutsPanel(it, settingsView.manager.globalConfig)
-					.redrawShortcuts())
-		}
-		tabs.minimumSize = Dimension(700, 750)
-		tabs.preferredSize = Dimension(700, 750)
-		tabs.maximumSize = Dimension(700, 750)
-		settingsView.rightPanel.removeAll()
-		settingsView.rightPanel.revalidate()
-		settingsView.rightPanel.add(tabs)
-		settingsView.rightPanel.repaint()
-	}
-
-	init
-	{
-		init()
-	}
-}
-
-/**
- * The [SettingPanelSelection] used for showing other settings.
- *
- * @author Richard Arriaga
- *
- * @constructor
- * Construct a [OtherSettingsSelection].
- *
- * @param settingsView
- *   The parent [SettingsView].
- */
-class OtherSettingsSelection constructor(
-	settingsView: SettingsView
-): SettingPanelSelection("Other", settingsView)
-{
-	override fun updateSettingsPane()
-	{
-		// TODO build screen
-		settingsView.rightPanel.removeAll()
-		settingsView.rightPanel.revalidate()
-		settingsView.rightPanel.add(Box.createRigidArea(Dimension(0, 70)))
-		settingsView.rightPanel.add(JLabel("WIP…"))
-		settingsView.rightPanel.repaint()
-	}
-
-	init
-	{
-		init()
-	}
-}
-
-/**
- * The [SettingPanelSelection] used for showing other settings.
- *
- * @author Richard Arriaga
- *
- * @constructor
- * Construct a [OtherSettingsSelection].
- *
- * @param settingsView
- *   The parent [SettingsView].
- */
-class StandardLibrariesSelection constructor(
-	settingsView: SettingsView
-): SettingPanelSelection("Standard Library", settingsView)
-{
-	override fun updateSettingsPane()
-	{
-		val panel = JPanel().apply {
-			layout = BoxLayout(this, BoxLayout.Y_AXIS)
-			minimumSize = Dimension(700, 800)
-			preferredSize = Dimension(700, 800)
-			maximumSize = Dimension(700, 800)
-			border = BorderFactory.createEmptyBorder(15, 10, 15, 10)
-		}
-		val libraries = availStandardLibraries
-		val latest =
-			this@StandardLibrariesSelection.settingsView.manager.latestVersion
-		if (latest.isNotEmpty())
-		{
-			val latestJarName = "avail-stdlib-$latest.jar"
-			val match = libraries.firstOrNull {
-				it.name == latestJarName
-			}
-			if (match == null)
-			{
-				panel.add(NewLibAvailableRow(latest))
-				panel.add(Box.createRigidArea(Dimension(0, 30)))
-			}
-		}
-		libraries.forEach { panel.add(LibRow(it)) }
-		settingsView.rightPanel.removeAll()
-		settingsView.rightPanel.revalidate()
-		settingsView.rightPanel.add(panel)
-		settingsView.rightPanel.add(JScrollPane().apply {
-			setViewportView(panel)
-			horizontalScrollBarPolicy =
-				ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER
-			verticalScrollBarPolicy =
-				ScrollPaneConstants.VERTICAL_SCROLLBAR_AS_NEEDED
-		})
-		settingsView.rightPanel.repaint()
-	}
-
-	init
-	{
-		init()
-	}
-
-	/**
-	 * A [JPanel] that displays an Avail standard library in
-	 * [AvailEnvironment.availHomeLibs].
-	 *
-	 * @property latest
-	 *   The latest available version
-	 */
-	inner class NewLibAvailableRow constructor(
-		private val latest: String
-	): JPanel(GridBagLayout())
-	{
-		/**
-		 * The [GridBagConstraints] used for all components in this
-		 * [NewLibAvailableRow].
-		 */
-		private val constraints = GridBagConstraints().apply {
-			anchor = GridBagConstraints.WEST
-		}
-
-		init
-		{
-			border = BorderFactory.createEmptyBorder()
-			minimumSize = Dimension(675, 35)
-			preferredSize = Dimension(675, 35)
-			maximumSize = Dimension(675, 35)
-
-			JPanel().apply {
-				border = BorderFactory.createEmptyBorder()
-				layout = BoxLayout(this, BoxLayout.X_AXIS)
-				add(Box.createRigidArea(Dimension(10, 0)))
-				add(JLabel("New Avail Standard Library Available: $latest")
-					.apply {
-						font = font.deriveFont(font.style or Font.ITALIC)
-					})
-				this@NewLibAvailableRow.add(
-					this,
-					constraints.apply {
-						weightx = 1.0
-					})
-			}
-
-			JButton("Install").apply {
-				addActionListener {
-					val selection = JOptionPane.showConfirmDialog(
-						this@NewLibAvailableRow,
-						"Install 'avail-stdlib-$latest.jar'?",
-						"Install Library",
-						JOptionPane.YES_NO_OPTION)
-					if (selection == 0)
-					{
-						SwingUtilities.invokeLater {
-							val byteCount =
-								MavenCentralAPI.downloadAvailStandardLib(latest)
-							if (byteCount == 0L)
-							{
-								System.err.println(
-									"Failed to download avail-stdlib: $latest")
-							}
-							this@StandardLibrariesSelection.updateSettingsPane()
-						}
-					}
-				}
-				this@NewLibAvailableRow.add(
-					this,
-					constraints.apply {
-						weightx = 1.0
-						anchor = GridBagConstraints.EAST
-					})
-			}
-		}
-	}
-
-	/**
-	 * A [JPanel] that displays an Avail standard library in
-	 * [AvailEnvironment.availHomeLibs].
-	 *
-	 * @property lib
-	 *   The library [File].
-	 */
-	inner class LibRow constructor(
-		private val lib: File
-	): JPanel(GridBagLayout())
-	{
-		/**
-		 * The [GridBagConstraints] used for all components in this [LibRow].
-		 */
-		private val constraints = GridBagConstraints().apply {
-			anchor = GridBagConstraints.WEST
-		}
-
-		init
-		{
-			border = BorderFactory.createEmptyBorder()
-			minimumSize = Dimension(675, 35)
-			preferredSize = Dimension(675, 35)
-			maximumSize = Dimension(675, 35)
-
-			JPanel().apply {
-				border = BorderFactory.createEmptyBorder()
-				layout = BoxLayout(this, BoxLayout.X_AXIS)
-				add(Box.createRigidArea(Dimension(10, 0)))
-				add(JLabel(lib.name))
-				this@LibRow.add(
-					this,
-					constraints.apply {
-						weightx = 1.0
-					})
-			}
-
-			JButton(deleteIcon).apply {
-				isContentAreaFilled = false
-				isBorderPainted = false
-				addActionListener {
-					val selection = JOptionPane.showConfirmDialog(
-						this@LibRow,
-						"Delete '${lib.name}'?",
-						"Delete Library",
-						JOptionPane.YES_NO_OPTION)
-					if (selection == 0)
-					{
-						lib.delete()
-						SwingUtilities.invokeLater {
-							this@StandardLibrariesSelection.updateSettingsPane()
-						}
-					}
-				}
-				this@LibRow.add(
-					this,
-					constraints.apply {
-						weightx = 1.0
-						anchor = GridBagConstraints.EAST
-					})
-			}
-		}
-	}
-
-	companion object
-	{
-		/**
-		 * The height to use for the favorite icons.
-		 */
-		private const val scaledIconHeight = 20
-
-		/**
-		 * The delete icon.
-		 */
-		private val deleteIcon get() =
-			ProjectManagerIcons.cancelFilled(scaledIconHeight)
-	}
-}
