+++ avail/src/main/kotlin/avail/anvil/settings/ShortcutsSelection.kt
@@ -1,0 +1,77 @@
+/*
+ * ShortcutsSelection.kt
+ * Copyright © 1993-2023, The Avail Foundation, LLC.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither the name of the copyright holder nor the names of the contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package avail.anvil.settings
+
+import avail.anvil.shortcuts.KeyboardShortcutCategory
+import java.awt.Dimension
+import javax.swing.JTabbedPane
+
+/**
+ * The [SettingPanelSelection] used for showing the shortcut settings.
+ *
+ * @author Richard Arriaga
+ *
+ * @constructor
+ * Construct a [ShortcutsSelection].
+ *
+ * @param settingsView
+ *   The parent [SettingsView].
+ */
+class ShortcutsSelection constructor(
+	settingsView: SettingsView
+): SettingPanelSelection("Shortcuts", settingsView)
+{
+	override fun updateSettingsPane()
+	{
+		val tabs = JTabbedPane(JTabbedPane.TOP, JTabbedPane.WRAP_TAB_LAYOUT)
+
+		KeyboardShortcutCategory.values().forEach {
+			tabs.addTab(
+				it.display,
+				ShortcutsPanel(it, settingsView.manager.globalConfig)
+					.redrawShortcuts())
+		}
+		tabs.minimumSize = Dimension(700, 750)
+		tabs.preferredSize = Dimension(700, 750)
+		tabs.maximumSize = Dimension(700, 750)
+		settingsView.rightPanel.removeAll()
+		settingsView.rightPanel.revalidate()
+		settingsView.rightPanel.add(tabs)
+		settingsView.rightPanel.repaint()
+	}
+
+	init
+	{
+		init()
+	}
+}
